Move rageshake configuration from resource file to Kotlin RageshakeConfig
