+++ .idea/dictionaries/shared.xml
@@ -13,6 +13,7 @@
             <w>onboarding</w>
             <w>placeables</w>
             <w>posthog</w>
+            <w>rageshake</w>
             <w>securebackup</w>
             <w>showkase</w>
             <w>snackbar</w>
