--- features/rageshake/impl/src/main/res/values/strings.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright (c) 2022 New Vector Ltd
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<resources>
-
-    <!-- Rageshake configuration -->
-    <string name="bug_report_url" translatable="false">https://riot.im/bugreports/submit</string>
-    <!--
-      As per https://github.com/matrix-org/rageshake
-      bug_report_app_name: Identifier for the application (eg 'riot-web').
-      Should correspond to a mapping configured in the configuration file for github issue reporting to work.
-    -->
-    <string name="bug_report_app_name" translatable="false">element-x-android</string>
-
-
-</resources>
